UI & UX update
